Standards: Added the STON language specification draft
Also, added a README file to navigate through the standards folder resources.
Also, rewrote language overview, mostly to account for language revisions, both in the nomenclature and the underlying concepts.

Useful suggestions to improve the draft are welcome, especially when it comes to making the specification more concise or accessible. I have little experience writing complete technical standards, and I have yet to grasp exactly what amount of precision is necessary. However, I plan to keep this specification text-only. If I make a STON specification with figures or diagrams (like the ones for JSON), I will publish it as another document.